Project Redirect - NetConnect Foundation
As I was unable to establish connection to the HTTP server from within the SFU network, I have moved away from the Kitura based HTTP server idea that was going to be the core of the network interface and have opted to create a custom API which I have tentatively called NetConnect. NetConnect uses the standard UDP network protocol to transmit messages from one host to another which is not by choice but necessity instead (it was the only protocol to work from within SFU's environment).

This commit contains the project directory structure along with the implementation of basic protocols and the Message class.

The directory server/Sources/NetConnect/ contains the custom network API, server/Sources/Application/ contains the main TrackITServer executable, and server/Tests/NetConnectTests/ contains unit tests for the NetConnect library.