Overview Page Upgrades, Information Page creation and bug fixes.
Made the charts work off of percentages, no more scroll view, updated the date label at the top. Also made a storyboard for the information on serving sizes and the dash diet but haven't edited the info yet. Also added some functions to time class. Finally I tried something to reset the values when the app detects it is being opened on a new day.